Added ollama support with model detection, provider menu and chat history. Moved to a simpler UI
