+++ terminal_chat/app/api/base.py
@@ -34,15 +34,40 @@ class BaseModelClient(ABC):
         from .anthropic import AnthropicClient
         from .openai import OpenAIClient
         
+        # For known models, use their configured provider
         model_info = CONFIG["available_models"].get(model_name)
-        if not model_info:
-            raise ValueError(f"Unknown model: {model_name}")
-            
-        provider = model_info["provider"]
+        if model_info:
+            provider = model_info["provider"]
+        else:
+            # For custom models, infer provider from name prefix
+            model_name_lower = model_name.lower()
+            if any(name in model_name_lower for name in ["gpt", "text-", "davinci"]):
+                provider = "openai"
+            elif any(name in model_name_lower for name in ["claude", "anthropic"]):
+                provider = "anthropic"
+            elif any(name in model_name_lower for name in ["llama", "mistral", "codellama", "gemma"]):
+                provider = "ollama"
+            else:
+                # Try to get from Ollama API first
+                from .ollama import OllamaClient
+                try:
+                    client = OllamaClient()
+                    models = client.get_available_models()
+                    if any(model["id"] == model_name for model in models):
+                        provider = "ollama"
+                    else:
+                        # Default to OpenAI if not found
+                        provider = "openai"
+                except:
+                    # Default to OpenAI if Ollama not available
+                    provider = "openai"
         
         if provider == "anthropic":
             return AnthropicClient()
         elif provider == "openai":
             return OpenAIClient()
+        elif provider == "ollama":
+            from .ollama import OllamaClient
+            return OllamaClient()
         else:
             raise ValueError(f"Unknown provider: {provider}")
