+++ terminal_chat/app/ui/model_selector.py
@@ -1,7 +1,7 @@
 from typing import Dict, List, Any, Optional
 from textual.app import ComposeResult
 from textual.containers import Container
-from textual.widgets import Select, Label
+from textual.widgets import Select, Label, Input
 from textual.widget import Widget
 from textual.message import Message
 
@@ -15,19 +15,38 @@ class ModelSelector(Container):
     ModelSelector {
         width: 100%;
         height: auto;
-        padding: 1;
+        padding: 0;
         background: $surface-darken-1;
     }
     
-    #model-label {
+    #selector-container {
         width: 100%;
-        text-align: left;
-        padding-bottom: 1;
+        layout: horizontal;
+        height: 3;
+        padding: 0;
     }
     
-    #model-select {
-        width: 100%;
+    #provider-select {
+        width: 30%;
         height: 3;
+        margin-right: 1;
+    }
+    
+    #model-select, #custom-model-input {
+        width: 1fr;
+        height: 3;
+    }
+
+    #custom-model-input {
+        display: none;
+    }
+
+    #custom-model-input.show {
+        display: block;
+    }
+
+    #model-select.hide {
+        display: none;
     }
     """
     
@@ -45,34 +64,109 @@ class ModelSelector(Container):
     ):
         super().__init__(name=name, id=id)
         self.selected_model = selected_model or CONFIG["default_model"]
+        # Handle custom models not in CONFIG
+        if self.selected_model in CONFIG["available_models"]:
+            self.selected_provider = CONFIG["available_models"][self.selected_model]["provider"]
+        else:
+            # Default to OpenAI for custom models
+            self.selected_provider = "openai"
         
     def compose(self) -> ComposeResult:
         """Set up the model selector"""
-        yield Label("Model:", id="model-label")
-        
-        # Get model options
-        options = []
-        for model_id, model_info in CONFIG["available_models"].items():
-            options.append((model_info["display_name"], model_id))
+        with Container(id="selector-container"):
+            # Provider options including Ollama
+            provider_options = [
+                ("OpenAI", "openai"),
+                ("Anthropic", "anthropic"),
+                ("Ollama", "ollama")
+            ]
             
-        yield Select(
-            options, 
-            id="model-select", 
-            value=self.selected_model, 
-            allow_blank=False
-        )
+            # Provider selector
+            yield Select(
+                provider_options,
+                id="provider-select",
+                value=self.selected_provider,
+                allow_blank=False
+            )
+            
+            # Model selector and custom input
+            is_custom = self.selected_model not in CONFIG["available_models"]
+            yield Select(
+                self._get_model_options(self.selected_provider) + [("Custom Model...", "custom")],
+                id="model-select",
+                value="custom" if is_custom else self.selected_model,
+                classes="hide" if is_custom else "",
+                allow_blank=False
+            )
+            yield Input(
+                value=self.selected_model if is_custom else "",
+                placeholder="Enter custom model name",
+                id="custom-model-input",
+                classes="" if is_custom else "hide"
+            )
+            
+    def _get_model_options(self, provider: str) -> List[tuple]:
+        """Get model options for a specific provider"""
+        options = [
+            (model_info["display_name"], model_id)
+            for model_id, model_info in CONFIG["available_models"].items()
+            if model_info["provider"] == provider
+        ]
+        
+        # Add available Ollama models
+        if provider == "ollama":
+            try:
+                from app.api.ollama import OllamaClient
+                ollama = OllamaClient()
+                ollama_models = ollama.get_available_models()
+                for model in ollama_models:
+                    if model["id"] not in CONFIG["available_models"]:
+                        options.append((model["name"], model["id"]))
+            except:
+                pass
+                
+        options.append(("Custom Model...", "custom"))
+        return options
         
     def on_select_changed(self, event: Select.Changed) -> None:
         """Handle select changes"""
-        if event.select.id == "model-select":
-            self.selected_model = event.value
-            self.post_message(self.ModelSelected(event.value))
-            # Return focus to chat input
-            chat_interface = self.app.query_one("#chat-interface", expect_type=ChatInterface)
-            if chat_interface:
-                input_field = chat_interface.query_one("#message-input")
-                if input_field:
-                    self.app.set_focus(input_field)
+        if event.select.id == "provider-select":
+            self.selected_provider = event.value
+            # Update model options
+            model_select = self.query_one("#model-select", Select)
+            model_options = self._get_model_options(self.selected_provider)
+            model_select.set_options(model_options)
+            # Select first model of new provider
+            if model_options:
+                self.selected_model = model_options[0][1]
+                model_select.value = self.selected_model
+                self.post_message(self.ModelSelected(self.selected_model))
+                
+        elif event.select.id == "model-select":
+            if event.value == "custom":
+                # Show custom input
+                model_select = self.query_one("#model-select")
+                custom_input = self.query_one("#custom-model-input")
+                model_select.add_class("hide")
+                custom_input.remove_class("hide")
+                custom_input.focus()
+            else:
+                # Hide custom input
+                model_select = self.query_one("#model-select")
+                custom_input = self.query_one("#custom-model-input")
+                model_select.remove_class("hide")
+                custom_input.add_class("hide")
+                self.selected_model = event.value
+                self.post_message(self.ModelSelected(event.value))
+
+    def on_input_changed(self, event: Input.Changed) -> None:
+        """Handle custom model input changes"""
+        if event.input.id == "custom-model-input":
+            value = event.value.strip()
+            if value:  # Only update if there's actual content
+                self.selected_model = value
+                self.post_message(self.ModelSelected(value))
+            
             
     def get_selected_model(self) -> str:
         """Get the current selected model ID"""
@@ -80,10 +174,20 @@ class ModelSelector(Container):
     
     def set_selected_model(self, model_id: str) -> None:
         """Set the selected model"""
+        self.selected_model = model_id
         if model_id in CONFIG["available_models"]:
-            self.selected_model = model_id
             select = self.query_one("#model-select", Select)
+            custom_input = self.query_one("#custom-model-input")
             select.value = model_id
+            select.remove_class("hide")
+            custom_input.add_class("hide")
+        else:
+            select = self.query_one("#model-select", Select)
+            custom_input = self.query_one("#custom-model-input")
+            select.value = "custom"
+            select.add_class("hide")
+            custom_input.value = model_id
+            custom_input.remove_class("hide")
 
 class StyleSelector(Container):
     """Widget for selecting the AI response style"""
@@ -92,18 +196,26 @@ class StyleSelector(Container):
     StyleSelector {
         width: 100%;
         height: auto;
-        padding: 1;
+        padding: 0;
         background: $surface-darken-1;
     }
     
-    #style-label {
+    #selector-container {
         width: 100%;
-        text-align: left;
-        padding-bottom: 1;
+        layout: horizontal;
+        height: 3;
+        padding: 0;
+    }
+    
+    #style-label {
+        width: 30%;
+        height: 3;
+        content-align: left middle;
+        padding-right: 1;
     }
     
     #style-select {
-        width: 100%;
+        width: 1fr;
         height: 3;
     }
     """
@@ -125,31 +237,26 @@ class StyleSelector(Container):
         
     def compose(self) -> ComposeResult:
         """Set up the style selector"""
-        yield Label("Response Style:", id="style-label")
-        
-        # Get style options
-        options = []
-        for style_id, style_info in CONFIG["user_styles"].items():
-            options.append((style_info["name"], style_id))
+        with Container(id="selector-container"):
+            yield Label("Style:", id="style-label")
             
-        yield Select(
-            options, 
-            id="style-select", 
-            value=self.selected_style, 
-            allow_blank=False
-        )
+            # Get style options
+            options = []
+            for style_id, style_info in CONFIG["user_styles"].items():
+                options.append((style_info["name"], style_id))
+                
+            yield Select(
+                options, 
+                id="style-select", 
+                value=self.selected_style, 
+                allow_blank=False
+            )
         
     def on_select_changed(self, event: Select.Changed) -> None:
         """Handle select changes"""
         if event.select.id == "style-select":
             self.selected_style = event.value
             self.post_message(self.StyleSelected(event.value))
-            # Return focus to chat input
-            chat_interface = self.app.query_one("#chat-interface", expect_type=ChatInterface)
-            if chat_interface:
-                input_field = chat_interface.query_one("#message-input")
-                if input_field:
-                    self.app.set_focus(input_field)
             
     def get_selected_style(self) -> str:
         """Get the current selected style ID"""
