+++ terminal_chat/legacy2.py
