+++ terminal_chat/main.py
@@ -1,326 +1,453 @@
-
 #!/usr/bin/env python3
+"""
+Simplified version of Terminal Chat with AI functionality
+"""
 import os
-import sys
-import asyncio
-import logging
-from typing import List, Dict, Any, Optional
-import time
-from pathlib import Path
+from typing import List, Optional, Callable
+from datetime import datetime
+
 from textual.app import App, ComposeResult
-from textual.containers import Container, Horizontal, Vertical
+from textual.containers import Container, Horizontal, Vertical, ScrollableContainer, Center
 from textual.reactive import reactive
-from textual.widgets import Header, Footer, Static
+from textual.widgets import Button, Input, Label, Static, Header, Footer, ListView, ListItem
 from textual.binding import Binding
 from textual import work
-
-from app.config import CONFIG, OPENAI_API_KEY, ANTHROPIC_API_KEY, APP_DIR
-
-# Set up logging
-log_path = APP_DIR / "debug.log"
-logging.basicConfig(
-    filename=str(log_path),
-    level=logging.DEBUG,
-    format='%(asctime)s - %(name)s - %(levelname)s - %(message)s'
-)
-logger = logging.getLogger('terminal_chat')
-
-# Also log to stderr for immediate feedback
-console_handler = logging.StreamHandler()
-console_handler.setLevel(logging.INFO)
-logger.addHandler(console_handler)
-from app.database import ChatDatabase
+from textual.screen import Screen
+from openai import OpenAI
 from app.models import Message, Conversation
-from app.api.base import BaseModelClient
-from app.ui.chat_interface import ChatInterface
-from app.ui.chat_list import ChatList
+from app.database import ChatDatabase
+from app.config import CONFIG, OPENAI_API_KEY, ANTHROPIC_API_KEY, OLLAMA_BASE_URL
+from app.ui.chat_interface import MessageDisplay
 from app.ui.model_selector import ModelSelector, StyleSelector
-from app.ui.search import SearchBar
-from app.ui.styles import CSS
-from app.utils import (
-    create_new_conversation, 
-    add_message_to_conversation,
-    update_conversation_title,
-    generate_streaming_response
-)
+from app.ui.chat_list import ChatList
+from app.api.base import BaseModelClient
+from app.utils import generate_streaming_response
 
-class TerminalChatApp(App):
-    """The main Terminal Chat application."""
+class SettingsScreen(Screen):
+    """Screen for model and style settings."""
     
-    TITLE = "Terminal Chat"
-    SUB_TITLE = "A ChatGPT/Claude experience in your terminal"
-    CSS = CSS + """
-    #main-layout {
-        width: 100%;
-        height: 100%;
-        layout: vertical;
+    CSS = """
+    #settings-container {
+        width: 60;
+        height: auto;
+        background: $surface;
+        border: round $primary;
+        padding: 1;
     }
     
+    #title {
+        width: 100%;
+        content-align: center middle;
+        text-align: center;
+        padding-bottom: 1;
+    }
+
+    #button-row {
+        width: 100%;
+        height: auto;
+        align-horizontal: right;
+        margin-top: 1;
+    }
+
+    #button-row Button {
+        width: auto;
+        min-width: 10;
+        margin-left: 1;
+    }
+    """
+
+    def compose(self) -> ComposeResult:
+        with Center():
+            with Container(id="settings-container"):
+                yield Static("Settings", id="title")
+                yield ModelSelector(self.app.selected_model)
+                yield StyleSelector(self.app.selected_style)
+                with Horizontal(id="button-row"):
+                    yield Button("Cancel", variant="default")
+                    yield Button("Done", variant="primary")
+
+    BINDINGS = [
+        Binding("escape", "action_cancel", "Cancel"),
+    ]
+
+    def action_cancel(self) -> None:
+        """Handle cancel action"""
+        self.app.pop_screen()
+
+    def on_button_pressed(self, event: Button.Pressed) -> None:
+        if event.button.label == "Done":
+            # Update current conversation with selected model and style
+            if self.app.current_conversation:
+                self.app.db.update_conversation(
+                    self.app.current_conversation.id,
+                    model=self.app.selected_model,
+                    style=self.app.selected_style
+                )
+                self.app.current_conversation.model = self.app.selected_model
+                self.app.current_conversation.style = self.app.selected_style
+            self.app.pop_screen()
+        elif event.button.label == "Cancel":
+            self.app.pop_screen()
+
+class HistoryScreen(Screen):
+    """Screen for viewing chat history."""
+    
+    BINDINGS = [
+        Binding("escape", "pop_screen", "Close"),
+    ]
+    
+    CSS = """
+    #history-container {
+        width: 80;
+        height: 40;
+        background: $surface;
+        border: round $primary;
+        padding: 1;
+    }
+    
+    #title {
+        width: 100%;
+        content-align: center middle;
+        text-align: center;
+        padding-bottom: 1;
+    }
+    
+    ListView {
+        width: 100%;
+        height: 1fr;
+        border: solid $primary;
+    }
+    
+    ListItem {
+        padding: 1;
+        border-bottom: solid $primary-darken-2;
+    }
+    
+    ListItem:hover {
+        background: $primary-darken-1;
+    }
+    
+    #button-row {
+        width: 100%;
+        height: 3;
+        align-horizontal: center;
+        margin-top: 1;
+    }
+    """
+
+    def __init__(self, conversations: List[dict], callback: Callable[[int], None]):
+        super().__init__()
+        self.conversations = conversations
+        self.callback = callback
+
+    def compose(self) -> ComposeResult:
+        with Center():
+            with Container(id="history-container"):
+                yield Static("Chat History", id="title")
+                
+                # Create list items for conversations
+                list_view = ListView()
+                for conv in self.conversations:
+                    title = conv["title"]
+                    model = conv["model"]
+                    if model in CONFIG["available_models"]:
+                        model = CONFIG["available_models"][model]["display_name"]
+                    item = ListItem(Label(f"{title} ({model})"))
+                    item.id = str(conv["id"])
+                    list_view.append(item)
+                yield list_view
+                
+                with Horizontal(id="button-row"):
+                    yield Button("Cancel", variant="primary")
+
+    def on_list_view_selected(self, event: ListView.Selected) -> None:
+        """Handle conversation selection."""
+        conv_id = int(event.item.id)
+        self.app.pop_screen()
+        self.callback(conv_id)
+
+    def on_button_pressed(self, event: Button.Pressed) -> None:
+        if event.button.label == "Cancel":
+            self.app.pop_screen()
+
+class SimpleChatApp(App):
+    """Simplified Terminal Chat application."""
+    
+    TITLE = "Simple Terminal Chat"
+    SUB_TITLE = "AI-powered chat"
+    
+    CSS = """
     #main-content {
-        width: 1fr;
+        width: 100%;
         height: 100%;
+        padding: 0 1;
+    }
+
+    #conversation-title {
+        width: 100%;
+        height: 3;
+        background: $primary-darken-1;
+        color: $text;
+        content-align: center middle;
+        text-align: center;
+    }
+
+    #messages-container {
+        width: 100%;
+        height: 1fr;
+        min-height: 10;
+        border-bottom: solid $primary-darken-2;
         overflow: auto;
         padding: 0 1;
     }
-    
-    #sidebar {
-        width: 30%;
-        min-width: 20;
-        max-width: 40%; 
-        height: 100%;
-        overflow: auto;
-    }
-    
-    #chat-interface {
+
+    #loading-indicator {
         width: 100%;
-        height: 100%;
+        height: 1;
+        background: $primary-darken-1;
+        color: $text;
+        content-align: center middle;
+        text-align: center;
+    }
+
+    #loading-indicator.hidden {
+        display: none;
+    }
+
+    #input-area {
+        width: 100%;
+        height: auto;
+        min-height: 4;
+        max-height: 10;
+        padding: 1;
+    }
+
+    #message-input {
+        width: 1fr;
+        min-height: 3;
+        height: auto;
+        margin-right: 1;
+        border: solid $primary-darken-2;
+    }
+
+    #message-input:focus {
+        border: tall $primary;
+    }
+
+    #send-button {
+        width: auto;
+        min-width: 10;
+        height: 3;
+    }
+
+    #button-row {
+        width: 100%;
+        height: auto;
+        align-horizontal: right;
+    }
+
+    #new-chat-button {
+        width: auto;
+        min-width: 10;
+        height: 3;
+        background: $success;
+    }
+
+    #view-history-button, #settings-button {
+        width: auto;
+        min-width: 10;
+        height: 3;
+        background: $primary;
+        margin-right: 1;
     }
     """
     
     BINDINGS = [
         Binding("q", "quit", "Quit"),
         Binding("n", "new_conversation", "New Chat"),
-        Binding("s", "toggle_sidebar", "Toggle Sidebar"),
-        Binding("f", "search", "Search"),
         Binding("escape", "escape", "Cancel"),
         Binding("ctrl+c", "quit", "Quit"),
     ]
     
     current_conversation = reactive(None)
-    sidebar_visible = reactive(True)
+    is_generating = reactive(False)
     
     def __init__(self):
-        try:
-            super().__init__()
-            self.db = ChatDatabase()
-            self.is_generating = False
-            logger.info("TerminalChatApp initialized")
-        except Exception as e:
-            logger.error(f"Error initializing app: {e}", exc_info=True)
-            raise
+        super().__init__()
+        self.db = ChatDatabase()
+        self.messages = []
+        self.selected_model = CONFIG["default_model"]
+        self.selected_style = CONFIG["default_style"]
         
     def compose(self) -> ComposeResult:
-        """Create the application layout."""
+        """Create the simplified application layout."""
         yield Header()
         
-        with Container(id="main-layout"):
-            with Horizontal():
-                # Sidebar
-                with Container(id="sidebar"):
-                    yield SearchBar(self.db, id="search-bar")
-                    yield ChatList(self.db, id="chat-list")
-                    with Container(id="model-container"):
-                        yield ModelSelector(id="model-selector")
-                        yield StyleSelector(id="style-selector")
-                
-                # Main content area
-                with Container(id="main-content"):
-                    yield ChatInterface(id="chat-interface")
+        with Vertical(id="main-content"):
+            # Conversation title
+            yield Static("New Conversation", id="conversation-title")
+            
+            # Messages area
+            with ScrollableContainer(id="messages-container"):
+                # Will be populated with messages
+                pass
+            
+            # Loading indicator
+            yield Static("Generating response...", id="loading-indicator", classes="hidden")
+            
+            # Input area
+            with Container(id="input-area"):
+                yield Input(placeholder="Type your message here...", id="message-input")
+                yield Button("Send", id="send-button", variant="primary")
+                with Horizontal(id="button-row"):
+                    yield Button("Settings", id="settings-button", variant="primary")
+                    yield Button("View History", id="view-history-button", variant="primary")
+                    yield Button("+ New Chat", id="new-chat-button")
         
         yield Footer()
         
     def on_mount(self) -> None:
         """Initialize the application on mount."""
+        # Check API keys and services
+        api_issues = []
+        if not OPENAI_API_KEY:
+            api_issues.append("- OPENAI_API_KEY is not set")
+        if not ANTHROPIC_API_KEY:
+            api_issues.append("- ANTHROPIC_API_KEY is not set")
+            
+        # Check Ollama availability
+        from app.api.ollama import OllamaClient
         try:
-            logger.info("App mounting")
-            # Apply initial layout
-            self.refresh_layout()
-            
-            # Check if API keys are set
-            api_issues = []
-            
-            if not OPENAI_API_KEY:
-                api_issues.append("- OPENAI_API_KEY is not set")
-                
-            if not ANTHROPIC_API_KEY:
-                api_issues.append("- ANTHROPIC_API_KEY is not set")
-                
-            if api_issues:
-                self.notify(
-                    "API key issues detected:\n" + "\n".join(api_issues) + 
-                    "\n\nSet them in your .env file or environment variables.",
-                    title="API Key Warning",
-                    severity="warning",
-                    timeout=10
-                )
-            
-            # Create a new conversation if none exists
-            self.create_new_conversation()
-            
-            # Focus the input field after layout is ready
-            self.call_after_refresh(self.focus_input)
-            
-        except Exception as e:
-            logger.error(f"Error during app mount: {e}", exc_info=True)
-            self.notify(f"Error initializing app: {str(e)}", severity="error")
+            ollama = OllamaClient()
+            models = ollama.get_available_models()
+            if not models:
+                api_issues.append("- No Ollama models found")
+        except Exception:
+            api_issues.append("- Ollama server not running")
         
-    def focus_input(self) -> None:
-        """Focus the input field"""
-        try:
-            logger.debug("Attempting to focus input")
-            chat_interface = self.query_one("#chat-interface", ChatInterface)
-            if chat_interface:
-                input_field = chat_interface.query_one("#message-input")
-                if input_field:
-                    logger.debug("Found input field, setting focus")
-                    # Clear any existing focus first
-                    if self.focused:
-                        self.focused.blur()
-                    # Set focus to input field
-                    input_field.focus()
-                    logger.debug("Focus set successfully")
-                else:
-                    logger.error("Input field not found")
-            else:
-                logger.error("Chat interface not found")
-        except Exception as e:
-            logger.error(f"Error focusing input: {e}", exc_info=True)
-            self.notify(f"Error focusing input: {str(e)}", severity="warning")
+        if api_issues:
+            self.notify(
+                "Service issues detected:\n" + "\n".join(api_issues) + 
+                "\n\nEnsure services are configured and running.",
+                title="Service Warning",
+                severity="warning",
+                timeout=10
+            )
+            
+        # Create a new conversation
+        self.create_new_conversation()
+        # Focus the input
+        self.query_one("#message-input").focus()
         
     def create_new_conversation(self) -> None:
         """Create a new chat conversation."""
-        # Get selected model and style with robust error handling
-        model = CONFIG["default_model"]
-        style = CONFIG["default_style"]
+        # Create new conversation in database using selected model and style
+        model = self.selected_model
+        style = self.selected_style
         
-        try:
-            model_selector = self.query_one("#model-selector", ModelSelector)
-            if model_selector:
-                model = model_selector.get_selected_model()
-        except Exception as e:
-            self.notify(f"Error getting model: {str(e)}", severity="warning")
-            
-        try:
-            style_selector = self.query_one("#style-selector", StyleSelector)
-            if style_selector:
-                style = style_selector.get_selected_style()
-        except Exception as e:
-            self.notify(f"Error getting style: {str(e)}", severity="warning")
+        # Create a title for the new conversation
+        title = f"New conversation ({datetime.now().strftime('%Y-%m-%d %H:%M')})"
         
-        try:
-            # Create new conversation in database
-            conversation = create_new_conversation(self.db, model, style)
-            self.current_conversation = conversation
-            
-            # Update UI safely
-            try:
-                chat_interface = self.query_one("#chat-interface", ChatInterface)
-                if chat_interface:
-                    chat_interface.set_conversation(conversation)
-            except Exception as e:
-                self.notify(f"Error updating chat interface: {str(e)}", severity="error")
-            
-            try:
-                chat_list = self.query_one("#chat-list", ChatList)
-                if chat_list:
-                    chat_list.refresh()
-                    chat_list.selected_id = conversation.id
-            except Exception as e:
-                self.notify(f"Error updating chat list: {str(e)}", severity="error")
-        except Exception as e:
-            self.notify(f"Failed to create conversation: {str(e)}", severity="error")
+        # Create conversation in database using the correct method
+        conversation_id = self.db.create_conversation(title, model, style)
+        
+        # Get the full conversation data
+        conversation_data = self.db.get_conversation(conversation_id)
+        
+        # Set as current conversation
+        self.current_conversation = Conversation.from_dict(conversation_data)
+        
+        # Update UI
+        title = self.query_one("#conversation-title", Static)
+        title.update(self.current_conversation.title)
+        
+        # Clear messages and update UI
+        self.messages = []
+        self.update_messages_ui()
         
     def action_new_conversation(self) -> None:
         """Handle the new conversation action."""
         self.create_new_conversation()
         
-    def action_toggle_sidebar(self) -> None:
-        """Toggle the sidebar visibility."""
-        self.sidebar_visible = not self.sidebar_visible
-        sidebar = self.query_one("#sidebar")
-        sidebar.display = True if self.sidebar_visible else False
-        
-    def action_search(self) -> None:
-        """Focus the search input."""
-        search_input = self.query_one("#search-bar #search-input")
-        self.set_focus(search_input)
-        
     def action_escape(self) -> None:
         """Handle escape key."""
         if self.is_generating:
-            # Stop generation
             self.is_generating = False
-            chat_interface = self.query_one("#chat-interface", ChatInterface)
-            chat_interface.stop_loading()
-            self.notify("Generation stopped.", severity="warning")
-            
-    def on_resize(self, event) -> None:
-        """Handle terminal resize events"""
-        try:
-            # Update the layout to fit the new terminal size
-            self.refresh_layout()
-            
-            # Re-focus the input field
-            chat_interface = self.query_one("#chat-interface", ChatInterface)
-            input_field = chat_interface.query_one("#message-input")
-            input_field.focus()
-        except Exception as e:
-            self.notify(f"Error handling resize: {str(e)}", severity="warning")
-            
-    def refresh_layout(self) -> None:
-        """Refresh the layout based on terminal size"""
-        try:
-            # Adjust sidebar width proportionally to terminal width
-            sidebar = self.query_one("#sidebar")
-            main_content = self.query_one("#main-content")
-            
-            # Set sidebar width based on terminal width
-            # Larger terminals get proportionally smaller sidebar
-            terminal_width = self.size.width
-            
-            if terminal_width < 80:  # Small terminal
-                sidebar_pct = 40
-            elif terminal_width < 120:  # Medium terminal
-                sidebar_pct = 35
-            else:  # Large terminal
-                sidebar_pct = 30
-                
-            sidebar.styles.width = f"{sidebar_pct}%"
-            main_content.styles.width = f"{100 - sidebar_pct}%"
-            
-            # Force a refresh of the rendering
-            self.refresh()
-        except Exception as e:
-            # Don't notify as this could be spammy
-            pass
+            self.notify("Generation stopped", severity="warning")
+            loading = self.query_one("#loading-indicator")
+            loading.add_class("hidden")
+        elif self.screen is not self.screen_stack[-1]:
+            # If we're in a sub-screen, pop it
+            self.pop_screen()
+    
+    def update_messages_ui(self) -> None:
+        """Update the messages UI."""
+        # Clear existing messages
+        messages_container = self.query_one("#messages-container")
+        messages_container.remove_children()
         
-    def on_chat_interface_message_sent(self, event: ChatInterface.MessageSent) -> None:
-        """Handle message sent from chat interface."""
-        if not self.current_conversation:
-            self.create_new_conversation()
+        # Add messages
+        for message in self.messages:
+            messages_container.mount(
+                MessageDisplay(message, highlight_code=CONFIG["highlight_code"])
+            )
             
-        # Add user message to conversation
-        add_message_to_conversation(
-            self.db, 
-            self.current_conversation, 
-            "user", 
-            event.content
+        # Scroll to bottom
+        messages_container.scroll_end(animate=False)
+    
+    def on_input_submitted(self, event: Input.Submitted) -> None:
+        """Handle input submission."""
+        self.action_send_message()
+    
+    def action_send_message(self) -> None:
+        """Initiate message sending."""
+        input_widget = self.query_one("#message-input", Input)
+        content = input_widget.value.strip()
+        
+        if not content or not self.current_conversation:
+            return
+        
+        # Clear input
+        input_widget.value = ""
+        
+        # Create user message
+        user_message = Message(role="user", content=content)
+        self.messages.append(user_message)
+        
+        # Save to database
+        self.db.add_message(
+            self.current_conversation.id,
+            "user",
+            content
         )
         
-        # Generate assistant response
-        self.generate_assistant_response()
+        # Update UI
+        self.update_messages_ui()
         
+        # Generate AI response
+        self.generate_response()
+        
+        # Focus back on input
+        input_widget.focus()
+    
     @work
-    async def generate_assistant_response(self) -> None:
-        """Generate an assistant response to the current conversation."""
-        if not self.current_conversation or not self.current_conversation.messages:
+    async def generate_response(self) -> None:
+        """Generate an AI response."""
+        if not self.current_conversation or not self.messages:
             return
             
         self.is_generating = True
-        chat_interface = self.query_one("#chat-interface", ChatInterface)
-        chat_interface.start_loading()
-        
-        # Ensure scrolled to bottom before generation starts
-        chat_interface.scroll_to_bottom()
+        loading = self.query_one("#loading-indicator")
+        loading.remove_class("hidden")
         
         try:
             # Get conversation parameters
-            model = self.current_conversation.model
-            style = self.current_conversation.style
+            model = self.selected_model
+            style = self.selected_style
             
             # Convert messages to API format
             api_messages = []
-            for msg in self.current_conversation.messages:
+            for msg in self.messages:
                 api_messages.append({
                     "role": msg.role,
                     "content": msg.content
@@ -331,6 +458,11 @@ class TerminalChatApp(App):
             
             # Start streaming response
             assistant_message = Message(role="assistant", content="")
+            self.messages.append(assistant_message)
+            messages_container = self.query_one("#messages-container")
+            message_display = MessageDisplay(assistant_message, highlight_code=CONFIG["highlight_code"])
+            messages_container.mount(message_display)
+            messages_container.scroll_end(animate=False)
             
             # Stream chunks to the UI
             async def update_ui(chunk: str):
@@ -339,7 +471,8 @@ class TerminalChatApp(App):
                 
                 try:
                     assistant_message.content += chunk
-                    chat_interface.add_message("assistant", assistant_message.content)
+                    message_display.update_content(assistant_message.content)
+                    messages_container.scroll_end(animate=False)
                 except Exception as e:
                     self.notify(f"Error updating UI: {str(e)}", severity="error")
                 
@@ -352,179 +485,81 @@ class TerminalChatApp(App):
                 update_ui
             )
             
-            # Add the full response to the database
+            # Save to database
             if self.is_generating:  # Only save if not cancelled
-                add_message_to_conversation(
-                    self.db,
-                    self.current_conversation,
+                self.db.add_message(
+                    self.current_conversation.id,
                     "assistant",
                     full_response
                 )
                 
-                # Update chat list
-                chat_list = self.query_one("#chat-list", ChatList)
-                chat_list.refresh()
-                
-                # Make sure we're scrolled to the bottom after response
-                chat_interface.scroll_to_bottom()
-                
-                # Re-focus the input field
-                input_field = chat_interface.query_one("#message-input")
-                input_field.focus()
-                
         except Exception as e:
             self.notify(f"Error generating response: {str(e)}", severity="error")
+            # Add error message
+            error_msg = f"Error generating response: {str(e)}"
+            self.messages.append(Message(role="assistant", content=error_msg))
+            self.update_messages_ui()
         finally:
             self.is_generating = False
-            chat_interface.stop_loading()
-            
-    def on_chat_list_new_chat_requested(self, event) -> None:
-        """Handle new chat request from the chat list."""
-        self.create_new_conversation()
-        
-    def on_chat_list_chat_selected(self, event: ChatList.ChatSelected) -> None:
-        """Handle chat selection from the chat list."""
-        try:
-            if not event or not event.conversation:
-                return
-                
-            self.current_conversation = event.conversation
-            
-            # Update UI with robust error handling
-            try:
-                chat_interface = self.query_one("#chat-interface", ChatInterface)
-                if chat_interface:
-                    chat_interface.set_conversation(event.conversation)
-            except Exception as e:
-                self.notify(f"Error updating chat interface: {str(e)}", severity="error")
-            
-            # Update model selector with robust error handling
-            try:
-                if event.conversation.model:
-                    model_selector = self.query_one("#model-selector", ModelSelector)
-                    if model_selector:
-                        model_selector.set_selected_model(event.conversation.model)
-            except Exception as e:
-                self.notify(f"Error updating model selector: {str(e)}", severity="warning")
-            
-            # Update style selector with robust error handling
-            try:
-                if event.conversation.style:
-                    style_selector = self.query_one("#style-selector", StyleSelector)
-                    if style_selector:
-                        style_selector.set_selected_style(event.conversation.style)
-            except Exception as e:
-                self.notify(f"Error updating style selector: {str(e)}", severity="warning")
-        except Exception as e:
-            self.notify(f"Error selecting chat: {str(e)}", severity="error")
-        
-    def on_search_bar_search_result_selected(self, event: SearchBar.SearchResultSelected) -> None:
-        """Handle search result selection."""
-        try:
-            if not event or not hasattr(event, 'conversation_id'):
-                return
-                
-            conversation_data = self.db.get_conversation(event.conversation_id)
-            if not conversation_data:
-                self.notify("Conversation not found", severity="warning")
-                return
-                
-            # Update chat list selection safely
-            try:
-                chat_list = self.query_one("#chat-list", ChatList)
-                if chat_list:
-                    chat_list.selected_id = event.conversation_id
-            except Exception as e:
-                self.notify(f"Error updating chat list: {str(e)}", severity="error")
-            
-            # Load conversation
-            try:
-                self.current_conversation = Conversation.from_dict(conversation_data)
-            except Exception as e:
-                self.notify(f"Error parsing conversation data: {str(e)}", severity="error")
-                return
-            
-            # Update UI components safely
-            try:
-                chat_interface = self.query_one("#chat-interface", ChatInterface)
-                if chat_interface:
-                    chat_interface.set_conversation(self.current_conversation)
-            except Exception as e:
-                self.notify(f"Error updating chat interface: {str(e)}", severity="error")
-            
-            # Update model selector safely
-            try:
-                if self.current_conversation.model:
-                    model_selector = self.query_one("#model-selector", ModelSelector)
-                    if model_selector:
-                        model_selector.set_selected_model(self.current_conversation.model)
-            except Exception as e:
-                self.notify(f"Error updating model selector: {str(e)}", severity="warning")
-            
-            # Update style selector safely
-            try:
-                if self.current_conversation.style:
-                    style_selector = self.query_one("#style-selector", StyleSelector)
-                    if style_selector:
-                        style_selector.set_selected_style(self.current_conversation.style)
-            except Exception as e:
-                self.notify(f"Error updating style selector: {str(e)}", severity="warning")
-        except Exception as e:
-            self.notify(f"Error handling search result: {str(e)}", severity="error")
+            loading = self.query_one("#loading-indicator")
+            loading.add_class("hidden")
             
     def on_model_selector_model_selected(self, event: ModelSelector.ModelSelected) -> None:
-        """Handle model selection."""
-        try:
-            if not event or not hasattr(event, 'model_id'):
-                return
-                
-            if not self.current_conversation:
-                self.notify("No active conversation to set model", severity="warning")
-                return
-                
-            model_id = event.model_id
-            
-            # Update model in the database with error handling
-            try:
-                self.db.update_conversation(self.current_conversation.id, model=model_id)
-                self.current_conversation.model = model_id
-                self.notify(f"Model changed to {model_id}", severity="information")
-            except Exception as e:
-                self.notify(f"Error updating model in database: {str(e)}", severity="error")
-        except Exception as e:
-            self.notify(f"Error handling model selection: {str(e)}", severity="error")
-            
+        """Handle model selection"""
+        self.selected_model = event.model_id
+        
     def on_style_selector_style_selected(self, event: StyleSelector.StyleSelected) -> None:
-        """Handle style selection."""
-        try:
-            if not event or not hasattr(event, 'style_id'):
+        """Handle style selection"""
+        self.selected_style = event.style_id
+        
+    def on_button_pressed(self, event: Button.Pressed) -> None:
+        """Handle button presses."""
+        button_id = event.button.id
+        
+        if button_id == "send-button":
+            self.action_send_message()
+        elif button_id == "new-chat-button":
+            self.create_new_conversation()
+        elif button_id == "settings-button":
+            self.push_screen(SettingsScreen())
+        elif button_id == "view-history-button":
+            self.view_chat_history()
+            
+    def view_chat_history(self) -> None:
+        """Show chat history in a popup."""
+        # Get recent conversations
+        conversations = self.db.get_all_conversations(limit=CONFIG["max_history_items"])
+        if not conversations:
+            self.notify("No chat history found", severity="warning")
+            return
+            
+        def handle_selection(selected_id: int) -> None:
+            if not selected_id:
                 return
                 
-            if not self.current_conversation:
-                self.notify("No active conversation to set style", severity="warning")
+            # Get full conversation
+            conversation_data = self.db.get_conversation(selected_id)
+            if not conversation_data:
+                self.notify("Could not load conversation", severity="error")
                 return
                 
-            style_id = event.style_id
+            # Update current conversation
+            self.current_conversation = Conversation.from_dict(conversation_data)
             
-            # Update style in the database with error handling
-            try:
-                self.db.update_conversation(self.current_conversation.id, style=style_id)
-                self.current_conversation.style = style_id
-                self.notify(f"Style changed to {style_id}", severity="information")
-            except Exception as e:
-                self.notify(f"Error updating style in database: {str(e)}", severity="error")
-        except Exception as e:
-            self.notify(f"Error handling style selection: {str(e)}", severity="error")
+            # Update title
+            title = self.query_one("#conversation-title", Static)
+            title.update(self.current_conversation.title)
             
-def main():
-    """Run the application."""
-    try:
-        logger.info("Starting Terminal Chat")
-        app = TerminalChatApp()
-        app.run()
-    except Exception as e:
-        logger.critical(f"Fatal error: {e}", exc_info=True)
-        raise
+            # Load messages
+            self.messages = [Message(**msg) for msg in self.current_conversation.messages]
+            self.update_messages_ui()
+            
+            # Update model and style selectors
+            self.selected_model = self.current_conversation.model
+            self.selected_style = self.current_conversation.style
+            
+        self.push_screen(HistoryScreen(conversations, handle_selection))
 
 if __name__ == "__main__":
-    main()
+    app = SimpleChatApp()
+    app.run()
