+++ terminal_chat/requirements.txt
@@ -2,5 +2,5 @@ textual>=0.11.1
 typer>=0.7.0
 requests>=2.28.1
 anthropic>=0.5.0
-openai>=0.27.0
+openai>=1.0.0
 python-dotenv>=0.21.0
